+++ app/public/wp-content/plugins/wp-reviews-plugin-for-google/tabs/setup_no_reg_header.php
@@ -13,8 +13,7 @@ $ti_command_list = [
 'save-align',
 'save-review-text-mode',
 'save-amp-notice-hide',
-'review-manual-download',
-'rate-us-feedback'
+'review-manual-download'
 ];
 if (!in_array($ti_command, $ti_command_list)) {
 $ti_command = null;
@@ -26,6 +25,9 @@ global $wpdb;
 if (!$pageDetails) {
 return false;
 }
+$trustindex_pm_google->setNotificationParam('not-using-no-connection', 'active', false);
+$trustindex_pm_google->setNotificationParam('not-using-no-widget', 'active', true);
+$trustindex_pm_google->setNotificationParam('not-using-no-widget', 'timestamp', time() + (2 * 3600));
 $tableName = $trustindex_pm_google->get_tablename('reviews');
 $wpdb->query('TRUNCATE `'. $tableName .'`');
 $reviews = null;
@@ -56,10 +58,12 @@ delete_option($trustindex_pm_google->get_option_name('review-download-token'));
 if ($reviewDownload) {
 update_option($trustindex_pm_google->get_option_name('review-download-inprogress'), $reviewDownload, false);
 update_option($trustindex_pm_google->get_option_name('review-manual-download'), $manualDownload, false);
+update_option($trustindex_pm_google->get_option_name('review-download-is-connecting'), 1, false);
 }
 else {
 delete_option($trustindex_pm_google->get_option_name('review-download-inprogress'));
 delete_option($trustindex_pm_google->get_option_name('review-manual-download'));
+delete_option($trustindex_pm_google->get_option_name('review-download-is-connecting'));
 }
 if (is_array($reviews)) {
 foreach ($reviews as $row) {
@@ -125,10 +129,14 @@ delete_option($trustindex_pm_google->get_option_name('verified-icon'));
 delete_option($trustindex_pm_google->get_option_name('enable-animation'));
 delete_option($trustindex_pm_google->get_option_name('show-arrows'));
 delete_option($trustindex_pm_google->get_option_name('show-header-button'));
+delete_option($trustindex_pm_google->get_option_name('reviews-load-more'));
 delete_option($trustindex_pm_google->get_option_name('show-reviewers-photo'));
 delete_option($trustindex_pm_google->get_option_name('widget-setted-up'));
 }
 $wpdb->query('TRUNCATE `'. $trustindex_pm_google->get_tablename('reviews') .'`');
+$trustindex_pm_google->setNotificationParam('not-using-no-connection', 'active', true);
+$trustindex_pm_google->setNotificationParam('not-using-no-connection', 'timestamp', time() + 86400);
+$trustindex_pm_google->setNotificationParam('not-using-no-widget', 'active', false);
 }
 function trustindex_plugin_change_step($step = 5)
 {
@@ -162,7 +170,7 @@ trustindex_plugin_disconnect_page();
 }
 }
 if ($ti_command === 'save-page') {
-check_admin_referer('save-noreg_' . $trustindex_pm_google->get_plugin_slug(), '_wpnonce_save');
+check_admin_referer('ti-save-page');
 $pageDetails = isset($_POST['page_details']) ? json_decode(stripcslashes($_POST['page_details']), true) : null;
 $reviewDownload = isset($_POST['review_download']) ? sanitize_text_field($_POST['review_download']) : 0;
 trustindex_plugin_connect_page($pageDetails, true, $reviewDownload);
@@ -170,11 +178,13 @@ header('Location: admin.php?page=' . sanitize_text_field($_GET['page']) .'&tab=s
 exit;
 }
 else if ($ti_command === 'delete-page') {
+check_admin_referer('ti-delete-page');
 trustindex_plugin_disconnect_page();
 header('Location: admin.php?page='. sanitize_text_field($_GET['page']) .'&tab=setup_no_reg');
 exit;
 }
 else if ($ti_command === 'save-style') {
+check_admin_referer('ti-save-style');
 $styleId = intval($_REQUEST['style_id']);
 update_option($trustindex_pm_google->get_option_name('style-id'), $styleId, false);
 delete_option($trustindex_pm_google->get_option_name('review-content'));
@@ -188,6 +198,7 @@ header('Location: admin.php?page='. sanitize_text_field($_GET['page']) .'&tab=se
 exit;
 }
 else if ($ti_command === 'save-set') {
+check_admin_referer('ti-save-set');
 update_option($trustindex_pm_google->get_option_name('scss-set'), sanitize_text_field($_REQUEST['set_id']), false);
 trustindex_plugin_change_step(4);
 $trustindex_pm_google->noreg_save_css(true);
@@ -197,24 +208,25 @@ header('Location: admin.php?page='. sanitize_text_field($_GET['page']) .'&tab=se
 exit;
 }
 else if ($ti_command === 'save-filter') {
+check_admin_referer('ti-save-filter');
 $filter = isset($_POST['filter']) ? sanitize_text_field($_POST['filter']) : null;
 $filter = json_decode(stripcslashes($filter), true);
 update_option($trustindex_pm_google->get_option_name('filter'), $filter, false);
 exit;
 }
 else if ($ti_command === 'save-language') {
-check_admin_referer('save-language_' . $trustindex_pm_google->get_plugin_slug(), '_wpnonce_language');
+check_admin_referer('ti-save-language');
 update_option($trustindex_pm_google->get_option_name('lang'), sanitize_text_field($_POST['lang']), false);
 delete_option($trustindex_pm_google->get_option_name('review-content'));
 exit;
 }
 else if ($ti_command === 'save-dateformat') {
-check_admin_referer('save-dateformat_'.$trustindex_pm_google->get_plugin_slug(), '_wpnonce_dateformat');
+check_admin_referer('ti-save-dateformat');
 update_option($trustindex_pm_google->get_option_name('dateformat'), sanitize_text_field($_POST['dateformat']), false);
 exit;
 }
 else if ($ti_command === 'save-options') {
-check_admin_referer('save-options_' . $trustindex_pm_google->get_plugin_slug(), '_wpnonce_options');
+check_admin_referer('ti-save-options');
 $r = 0;
 if (isset($_POST['verified-icon'])) {
 $r = sanitize_text_field($_POST['verified-icon']);
@@ -235,6 +247,11 @@ if (isset($_POST['show-header-button'])) {
 $r = sanitize_text_field($_POST['show-header-button']);
 }
 update_option($trustindex_pm_google->get_option_name('show-header-button'), $r, false);
+$r = 1;
+if (isset($_POST['reviews-load-more'])) {
+$r = sanitize_text_field($_POST['reviews-load-more']);
+}
+update_option($trustindex_pm_google->get_option_name('reviews-load-more'), $r, false);
 $r = 1;
 if (isset($_POST['show-reviewers-photo'])) {
 $r = sanitize_text_field($_POST['show-reviewers-photo']);
@@ -270,13 +287,13 @@ $trustindex_pm_google->noreg_save_css(true);
 exit;
 }
 else if ($ti_command === 'save-align') {
-check_admin_referer('save-align_' . $trustindex_pm_google->get_plugin_slug(), '_wpnonce_align');
+check_admin_referer('ti-save-align');
 update_option($trustindex_pm_google->get_option_name('align'), sanitize_text_field($_POST['align']), false);
 $trustindex_pm_google->noreg_save_css(true);
 exit;
 }
 else if ($ti_command === 'save-review-text-mode') {
-check_admin_referer('save-review-text-mode_' . $trustindex_pm_google->get_plugin_slug(), '_wpnonce_review_text_mode');
+check_admin_referer('ti-save-review-text-mode');
 update_option($trustindex_pm_google->get_option_name('review-text-mode'), sanitize_text_field($_POST['review_text_mode']), false);
 $trustindex_pm_google->noreg_save_css(true);
 exit;
@@ -286,6 +303,7 @@ update_option($trustindex_pm_google->get_option_name('amp-hidden-notification'),
 exit;
 }
 else if ($ti_command === 'review-manual-download') {
+check_admin_referer('ti-download-reviews');
 $response = wp_remote_post('https://admin.trustindex.io/source/wordpressPageRequest', [
 'body' => [ 'id' => get_option($trustindex_pm_google->get_option_name('review-download-request-id')) ],
 'timeout' => '30',
@@ -316,21 +334,17 @@ status_header(404);
 }
 exit;
 }
-else if ($ti_command === 'rate-us-feedback') {
-$text = isset($_POST['text']) ? trim(wp_kses_post(stripslashes($_POST['text']))) : "";
-$email = isset($_POST['email']) ? trim(sanitize_text_field($_POST['email'])) : "";
-$star = isset($_REQUEST['star']) ? intval($_REQUEST['star']) : 1;
-update_option($trustindex_pm_google->get_option_name('rate-us-feedback'), $star, false);
-if ($star > 3) {
-header('Location: https://wordpress.org/support/plugin/'. $trustindex_pm_google->get_plugin_slug() . '/reviews/?rate='. $star .'#new-post');
-}
-else {
-wp_mail('support@trustindex.io', 'Feedback from Google plugin', "We received a <strong>$star star</strong> feedback about the Google plugin from $email:<br /><br />$text", [
-'From: '. $email,
-'Content-Type: text/html; charset=UTF-8'
-]);
-}
+if (isset($_GET['recreate'])) {
+check_admin_referer('ti-recreate');
+$trustindex_pm_google->uninstall();
+$trustindex_pm_google->activate();
+header('Location: admin.php?page=' . sanitize_text_field($_GET['page']) .'&tab=setup_no_reg');
 exit;
+}
+if (isset($_GET['setup_widget'])) {
+check_admin_referer('ti-setup-widget');
+update_option($trustindex_pm_google->get_option_name('widget-setted-up'), 1, false);
+header('Location: admin.php?page=' . sanitize_text_field($_GET['page']) .'&tab=setup_no_reg');
 }
 $reviews = [];
 if ($trustindex_pm_google->is_noreg_linked()) {
@@ -347,26 +361,16 @@ $verifiedIcon = get_option($trustindex_pm_google->get_option_name('verified-icon
 $enableAnimation = get_option($trustindex_pm_google->get_option_name('enable-animation'), 1);
 $showArrows = get_option($trustindex_pm_google->get_option_name('show-arrows'), 1);
 $showHeaderButton = get_option($trustindex_pm_google->get_option_name('show-header-button'), 1);
+$reviewsLoadMore = get_option($trustindex_pm_google->get_option_name('reviews-load-more'), 1);
 $widgetSettedUp = get_option($trustindex_pm_google->get_option_name('widget-setted-up'), 0);
 $disableFont = get_option($trustindex_pm_google->get_option_name('disable-font'), 0);
 $align = get_option($trustindex_pm_google->get_option_name('align'), in_array($styleId, [ 36, 37, 38, 39 ]) ? 'center' : 'left');
 $reviewTextMode = get_option($trustindex_pm_google->get_option_name('review-text-mode'), 'readmore');
-$rateUsFeedback = get_option($trustindex_pm_google->get_option_name('rate-us-feedback'), 0);
 $footerFilterText = get_option($trustindex_pm_google->get_option_name('footer-filter-text'), 0);
 $scssSetTmp = $scssSet ? $scssSet : 'light-background';
 $showReviewersPhoto = get_option($trustindex_pm_google->get_option_name('show-reviewers-photo'), TrustindexPlugin_google::$widget_styles[ $scssSetTmp ]['reviewer-photo'] ? 1 : 0);
 $showLogos = get_option($trustindex_pm_google->get_option_name('show-logos'), TrustindexPlugin_google::$widget_styles[ $scssSetTmp ]['hide-logos'] ? 0 : 1);
 $showStars = get_option($trustindex_pm_google->get_option_name('show-stars'), TrustindexPlugin_google::$widget_styles[ $scssSetTmp ]['hide-stars'] ? 0 : 1);
-if (isset($_GET['recreate'])) {
-$trustindex_pm_google->uninstall();
-$trustindex_pm_google->activate();
-header('Location: admin.php?page=' . sanitize_text_field($_GET['page']) .'&tab=setup_no_reg');
-exit;
-}
-if (isset($_GET['setup_widget'])) {
-update_option($trustindex_pm_google->get_option_name('widget-setted-up'), 1, false);
-$widgetSettedUp = 1;
-}
 $currentStep = isset($_GET['step']) ? intval(sanitize_text_field($_GET['step'])) : 0;
 if ($currentStep === 3 && in_array($styleId, [ 17, 21, 52, 53 ])) {
 $currentStep = 4;
